Fixed sidebar animations and scrolling
Scrollbars now show an X icon when opened (and it is visible). The
sidebar animations were also changed so that they shift onto the screen,
allowing text position in sidebar to remain the same. For desktop, the
scrolling onto different pages also highlights the tab you're currently
on instead of which one was clicked. Need to update this for mobile