Fixed bugs and updated projects tabs
Fixed a bug with the disappearing project menu (timeouts weren't cleared
with project tab selected, so it would trigger mouseleave setTimeout
when project tab selected). Also updated the project tabs (so there are
5 and properly labelled), just need to actually fill them with relevant
info. Also, for mobile devices, the nav menu is hidden after a tab is
selected (so tabs don't have to be updated when scrolling on mobile).