Bunch of fixes
Made the tools container arrows only appear if it can be scrolled. Also
fixed a bug with the tools logos not being properly centered on
Microsoft Edge and the padding not being on top of the projects picture
of me. Finished styling the buttons on the bottom container of the
projects, and now starting to added formatting for the hobbies